feat: move AuthIcons, rename providers to loginProviders, improve type safety
